requirements.txt
streamlit
langchain
langchain-community
langchain-openai
langchain-text-splitters
chromadb
pypdf
python-dotenv
sentence-transformers